Carry a TINYINT's sign out to the provider
Java's byte is signed and IKVM's is not, so java.lang.Byte.byteValue()
answers a CLR byte holding the two's complement bits: a Calcite TINYINT
of -56 bound to the provider as 200. AdoReaderUtil.GetByte already reads
the other direction correctly, through an sbyte.

An sbyte cannot be the answer on the way out -- SqlClient refuses the
type, "The parameter data type of SByte is invalid" -- so the value is
sign-extended to a short, the widening the unsigned arms beside it
already do for the types no provider binds.

A real backend reaches the arm only through a cast, SQL Server's own
tinyint being unsigned and a UTINYINT. The correlation test states the
type with a VALUES row and compares it against ID on the server, where
-56 and 200 fall either side; it fails on all three drivers without the
fix. AdoParameterValueTests reads the bound value itself and pins the
neighbouring arms, which are right as they stand.

+++ src/Apache.Calcite.Adapter.AdoNet/AdoEnumerable.cs
@@ -258,7 +258,12 @@ namespace Apache.Calcite.Adapter.AdoNet
             {
                 null => null,
                 java.lang.Boolean b => b.booleanValue(),
-                java.lang.Byte b => b.byteValue(),
+                // Java's byte is signed and IKVM's byte is not, so byteValue() answers a CLR byte
+                // holding the two's complement bits, and a TINYINT of -56 would bind as 200.
+                // shortValue() sign-extends instead, and short is the narrowest CLR type every
+                // provider binds: SqlClient refuses an sbyte outright, "The parameter data type of
+                // SByte is invalid", the same wall the unsigned types below run into
+                java.lang.Byte b => b.shortValue(),
                 java.lang.Short s => s.shortValue(),
                 java.lang.Integer i => i.intValue(),
                 java.lang.Long l => l.longValue(),
